# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ResuSight API",
    "description": "API documentation for ResuSight",
    "version": "1.0"
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/info/{username}": {
      "get": {
        "summary": "Get user details",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "example": {
                  "user_details": {
                    "username": "example_user",
                    "name": "John Doe",
                    "email": "john@example.com"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/generate/{username}/resume/{section}": {
      "get": {
        "summary": "Generate questions for a user's resume",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "section",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "example": {
                  "Questions": ["q1", "q2", "q3"]
                }
              }
            }
          }
        }
      }
    },
    "/generate/jd": {
      "post": {
        "summary": "Generate questions for a job description",
        "requestBody": {
          "required": true,
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "example": {
                  "Response": ["question1", "question2", "question3"]
                }
              }
            }
          }
        }
      }
    },
    "/register": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "The username of the user to register."
                  },
                  "email": {
                    "type": "string",
                    "description": "The email of the user to register."
                  },
                  "password": {
                    "type": "string",
                    "description": "The password of the user to register."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registration Successful"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Login with username and password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "The username of the user."
                  },
                  "password": {
                    "type": "string",
                    "description": "The password of the user."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login Successful"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Invalid Credentials"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/logout": {
      "post": {
        "summary": "Logout",
        "responses": {
          "200": {
            "description": "Logout successful"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "message": "No user data found"
                }
              }
            }
          }
        }
      }
    }
  }
}